requirements.txt
pandas
mysql-connector-python
neo4j
PyYAML
pyarrow
fastparquet